Geofenceのエントリ時のAPI通信をRepositoryパターンに変更
