+++ app/src/main/java/com/example/fitbattleandroid/repository/GeofenceEntryRepository.kt
@@ -1,0 +1,8 @@
+package com.example.fitbattleandroid.repository
+
+import com.example.fitbattleandroid.data.remote.EntryGeoFenceReq
+import com.example.fitbattleandroid.data.remote.EntryGeoFenceRes
+
+interface GeofenceEntryRepository {
+    suspend fun sendGeofenceEntryRequest(entryGeofenceReq: EntryGeoFenceReq): EntryGeoFenceRes
+}
