+++ app/src/main/java/com/example/fitbattleandroid/ui/navigation/NavHost.kt
@@ -27,6 +27,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
+import com.example.fitbattleandroid.data.EncounterRemoteDatasource
+import com.example.fitbattleandroid.repositoryImpl.GeofenceEntryRepositoryImpl
 import com.example.fitbattleandroid.ui.screen.EncounterHistoryScreen
 import com.example.fitbattleandroid.ui.screen.FitnessMemory
 import com.example.fitbattleandroid.ui.screen.LoginScreen
@@ -83,7 +85,10 @@ fun App(
             MainNavigation(
                 requestPermissionLauncher,
                 mapViewModel,
-                dataAPIViewModel = viewModel(),
+                dataAPIViewModel =
+                    viewModel {
+                        HealthDataApiViewModel(GeofenceEntryRepositoryImpl(EncounterRemoteDatasource()))
+                    },
                 backgroundPermissionGranted,
                 healthConnectClient,
             )
