+++ app/src/main/java/com/example/fitbattleandroid/viewmodel/HealthDataApiViewModel.kt
@@ -4,56 +4,23 @@ import androidx.lifecycle.ViewModel
 import com.example.fitbattleandroid.data.remote.EntryGeoFenceReq
 import com.example.fitbattleandroid.data.remote.EntryGeoFenceRes
 import com.example.fitbattleandroid.data.remote.MemberInfo
-import com.example.fitbattleandroid.data.remote.SaveFitnessReq
-import io.ktor.client.HttpClient
-import io.ktor.client.call.body
-import io.ktor.client.plugins.contentnegotiation.ContentNegotiation
-import io.ktor.client.request.post
-import io.ktor.client.request.setBody
-import io.ktor.client.statement.HttpResponse
-import io.ktor.http.ContentType
-import io.ktor.http.contentType
-import io.ktor.serialization.kotlinx.json.json
+import com.example.fitbattleandroid.repositoryImpl.GeofenceEntryRepositoryImpl
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.serialization.json.Json
 
-class HealthDataApiViewModel : ViewModel() {
+class HealthDataApiViewModel(
+    private val repository: GeofenceEntryRepositoryImpl,
+) : ViewModel() {
     private val _encounterMembers = MutableStateFlow<List<MemberInfo>>(emptyList())
     val encounterMembers: StateFlow<List<MemberInfo>> = _encounterMembers.asStateFlow()
 
-    private val client =
-        HttpClient {
-            install(ContentNegotiation) {
-                json(
-                    Json {
-                        ignoreUnknownKeys = true
-                        isLenient = true
-                    },
-                )
-            }
-        }
-
     // ジオフェンス入室リクエスト関数
     suspend fun sendGeoFenceEntryRequest(entry: EntryGeoFenceReq): EntryGeoFenceRes {
-        val res =
-            client.post("http://192.168.224.234:7070/api/v1/geofence/entry") {
-                contentType(ContentType.Application.Json)
-                setBody(entry)
-            }
-        // レスポンスからボディを取得して変数に追加
-        val responseBody = res.body<EntryGeoFenceRes>()
-        for (member in responseBody.passingMember) {
+        val res = repository.sendGeofenceEntryRequest(entry)
+        for (member in res.passingMember) {
             _encounterMembers.value += member
         }
-        return responseBody
+        return res
     }
-
-    // フィットネスデータの保存リクエスト
-    suspend fun sendFitnessSave(request: SaveFitnessReq): HttpResponse =
-        client.post("http://192.168.224.234:7070/api/v1/fitness/save") {
-            contentType(ContentType.Application.Json)
-            setBody(request)
-        }
 }
